pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tools-automation"
version = "0.1.0"
description = "Tools automation system"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.13.2",
    "aiosignal==1.4.0",
    "annotated-types==0.7.0",
    "anyio==4.11.0",
    "application_properties==0.9.0",
    "attrs==25.4.0",
    "blinker==1.9.0",
    "certifi==2025.10.5",
    "charset-normalizer==3.4.4",
    "click==8.3.0",
    "Columnar==1.4.1",
    "contourpy==1.3.3",
    "coverage==7.11.2",
    "cycler==0.12.1",
    "execnet==2.1.1",
    "filelock==3.20.0",
    "Flask==3.1.2",
    "flask-cors==6.0.1",
    "fonttools==4.60.1",
    "frozenlist==1.8.0",
    "fsspec==2025.10.0",
    "h11==0.16.0",
    "hf-xet==1.2.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "huggingface-hub==0.36.0",
    "idna==3.11",
    "iniconfig==2.3.0",
    "itsdangerous==2.2.0",
    "Jinja2==3.1.6",
    "joblib==1.5.2",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "kiwisolver==1.4.9",
    "langchain==1.0.5",
    "langchain-core==1.0.4",
    "langgraph==1.0.3",
    "langgraph-checkpoint==3.0.1",
    "langgraph-prebuilt==1.0.2",
    "langgraph-sdk==0.2.9",
    "langsmith==0.4.42",
    "MarkupSafe==3.0.3",
    "matplotlib==3.10.7",
    "multidict==6.7.0",
    "numpy==2.3.4",
    "orjson==3.11.4",
    "ormsgpack==1.12.0",
    "packaging==25.0",
    "pandas==2.3.3",
    "pillow==12.0.0",
    "pluggy==1.6.0",
    "prometheus_client==0.23.1",
    "propcache==0.4.1",
    "psutil==7.1.3",
    "pydantic==2.12.4",
    "pydantic_core==2.41.5",
    "Pygments==2.19.2",
    "pyjson5==2.0.0",
    "PyJWT==2.10.1",
    "pymarkdownlnt==0.9.33",
    "pyparsing==3.2.5",
    "pytest==8.4.2",
    "pytest-asyncio==1.3.0",
    "pytest-cov==7.0.0",
    "pytest-timeout==2.4.0",
    "pytest-xdist==3.8.0",
    "python-dateutil==2.9.0.post0",
    "pytz==2025.2",
    "PyYAML==6.0.3",
    "redis==7.0.1",
    "regex==2025.11.3",
    "requests==2.32.5",
    "requests-toolbelt==1.0.0",
    "safetensors==0.6.2",
    "scikit-learn==1.7.2",
    "scipy==1.16.3",
    "seaborn==0.13.2",
    "six==1.17.0",
    "sniffio==1.3.1",
    "tenacity==9.1.2",
    "threadpoolctl==3.6.0",
    "tokenizers==0.22.1",
    "tomli==2.3.0",
    "toolz==1.1.0",
    "tqdm==4.67.1",
    "transformers==4.57.1",
    "typing-inspection==0.4.2",
    "typing_extensions==4.15.0",
    "tzdata==2025.2",
    "urllib3==2.5.0",
    "wcwidth==0.2.14",
    "Werkzeug==3.1.2",
    "xxhash==3.6.0",
    "yarl==1.22.0",
    "zstandard==0.25.0"
]

[tool.setuptools]
packages = ["agents", "scripts", "tools"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --cov=agents --cov=scripts --cov-report=html --cov-report=term"

[tool.coverage.run]
source = ["agents", "scripts", "tools"]
omit = ["*/tests/*", "*/test_*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError"
]

[tool.ruff]
line-length = 100
extend-ignore = ["E203"]
exclude = [
    "venv",
    ".venv",
    "community/showcase",
    "AvoidObstaclesGame",
    "HabitQuest",
    "MomentumFinance",
    "PlannerApp",
    "shared-kit",
    "site",
    "reports",
    "sdk/python/docs",
    "sdk/python/examples",
    "docs",
]
